requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
filelock==3.24.0
Flask==3.1.2
fsspec==2026.2.0
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pillow==12.1.1
Pygments==2.19.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
shellingham==1.5.4
torch==2.10.0
tqdm==4.67.3
typer==0.23.1
typer-slim==0.23.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
